relax predicate to also check non unicode numerics
As a consequence of putraw(), not everything is a string. This meant that numeric objects (like integers, floats,...) are possible; not only strings. In order to only (re)format numerics that have been formatted as strings already, the type check was added. This commit makes sure that also non unicode strings with numerics in them are formatted.